resources/requirements.txt
imutils==0.5.4
opencv_python==4.8.0.74
numpy==1.21.5
pyserial==3.5
